SocketApiGeneral/pyproject.toml
[tool.poetry]
name = "apigeneral"
version = "0.1"
description = ""
authors = ["ramses314"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.1"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy-utils = "^0.41.2"
sqlalchemy = {version = "2.0.29", extras = ["all"]}
asyncpg = "^0.29.0"
python-socketio = "^5.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
